chore: restructure shared folder
