+++ frontend/lib/features/home/transaction_details_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_starter/l10n/app_localizations.dart';
-import 'package:flutter_starter/shared/grid.dart';
+import 'package:flutter_starter/shared/theme/grid.dart';
 import 'package:flutter_starter/shared/success_page.dart';
-import 'package:flutter_starter/shared/transaction.dart';
+import 'package:flutter_starter/features/home/transaction.dart';
 
 class TransactionDetailsPage extends HookWidget {
   final Transaction txn;
