+++ frontend/lib/features/payments/search_payment_methods_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:flutter_starter/features/payments/payment_method.dart';
 import 'package:flutter_starter/l10n/app_localizations.dart';
-import 'package:flutter_starter/shared/grid.dart';
-import 'package:flutter_starter/shared/payment_method.dart';
+import 'package:flutter_starter/shared/theme/grid.dart';
 
 class SearchPaymentMethodsPage extends HookWidget {
   final _formKey = GlobalKey<FormState>();
