+++ frontend/lib/features/pfis/pfis_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_starter/features/pfis/pfi_verification_page.dart';
 import 'package:flutter_starter/features/pfis/pfis_notifier.dart';
 import 'package:flutter_starter/l10n/app_localizations.dart';
-import 'package:flutter_starter/shared/grid.dart';
+import 'package:flutter_starter/shared/theme/grid.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class PfisPage extends HookConsumerWidget {
