+++ frontend/lib/features/send/send_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_starter/features/send/send_did_page.dart';
 import 'package:flutter_starter/l10n/app_localizations.dart';
 import 'package:flutter_starter/shared/number_pad.dart';
-import 'package:flutter_starter/shared/grid.dart';
+import 'package:flutter_starter/shared/theme/grid.dart';
 import 'package:flutter_starter/shared/animations/invalid_number_pad_input_animation.dart';
 import 'package:flutter_starter/shared/utils/number_pad_input_validation_util.dart';
