+++ frontend/lib/shared/animations/invalid_number_pad_input_animation.dart
@@ -34,6 +34,7 @@ class InvalidNumberPadInputAnimation extends HookWidget {
           });
         }
       });
+      return;
     }, []);
 
     useEffect(() {
