+++ frontend/lib/shared/json_schema_form.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_starter/l10n/app_localizations.dart';
-import 'package:flutter_starter/shared/grid.dart';
+import 'package:flutter_starter/shared/theme/grid.dart';
 
 class JsonSchemaForm extends HookWidget {
   final String schema;
