+++ frontend/test/features/home/home_page_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_starter/features/deposit/deposit_page.dart';
 import 'package:flutter_starter/features/home/home_page.dart';
 import 'package:flutter_starter/features/withdraw/withdraw_page.dart';
-import 'package:flutter_starter/shared/transaction.dart';
+import 'package:flutter_starter/features/home/transaction.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../helpers/widget_helpers.dart';
