+++ frontend/test/features/home/transaction_details_page_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_starter/features/home/transaction_details_page.dart';
-import 'package:flutter_starter/shared/transaction.dart';
+import 'package:flutter_starter/features/home/transaction.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../../helpers/widget_helpers.dart';
