+++ frontend/test/features/payments/payment_details_page_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_starter/shared/payment_details_page.dart';
-import 'package:flutter_starter/shared/payment_method.dart';
-import 'package:flutter_starter/shared/search_payment_methods_page.dart';
+import 'package:flutter_starter/features/payments/payment_details_page.dart';
+import 'package:flutter_starter/features/payments/payment_method.dart';
+import 'package:flutter_starter/features/payments/search_payment_methods_page.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../helpers/widget_helpers.dart';
+import '../../helpers/widget_helpers.dart';
 
 void main() {
   group('PaymentDetailsPage', () {
