+++ frontend/test/features/payments/search_payment_methods_page_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_starter/shared/payment_method.dart';
-import 'package:flutter_starter/shared/search_payment_methods_page.dart';
+import 'package:flutter_starter/features/payments/payment_method.dart';
+import 'package:flutter_starter/features/payments/search_payment_methods_page.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../helpers/widget_helpers.dart';
+import '../../helpers/widget_helpers.dart';
 
 final _paymentMethods = [
   PaymentMethod(
